requirements.txt
Flask
connexion
Flask-Cache
pyweatherbit
requests
redis
